change replacements in controllers

+++ backend/src/controllers/bookController.js
@@ -3,94 +3,97 @@ const db = require("../config/db");
 
 const BookController = {
     async getWantedBooks(req, res) {
-        const { user_id } = req.body;
+        const { user_id } = req.params.user_id;
         try {
             const wanted_books = await db.query(
-                "SELECT \
-                    w.book_id, \
-                    b.isbn, \
-                    gbi.title, \
-                    gbi.author_name, \
-                    gbi.category_name, \
-                    gbi.publish_date, \
-                    gbi.language, \
-                    gbi.series_name \
-                FROM wanted w \
-                LEFT JOIN book b ON b.book_id = w.book_id \
-                LEFT JOIN LATERAL get_book_info(b.isbn) gbi ON TRUE \
-                WHERE w.user_id = ?",
+                `SELECT
+                    w.book_id,
+                    b.isbn,
+                    gbi.title,
+                    gbi.author_name,
+                    gbi.category_name,
+                    gbi.publish_date,
+                    gbi.language,
+                    gbi.series_name
+                FROM wanted w
+                LEFT JOIN book b ON b.book_id = w.book_id
+                LEFT JOIN LATERAL get_book_info(b.isbn) gbi ON TRUE
+                WHERE w.user_id = :user_id
+                `,
                 {
                     type: QueryTypes.SELECT,
-                    replacements: [user_id]
+                    replacements: { user_id: user_id }
                 }
             );
             res.status(200).json(wanted_books);
         } catch (error) {
             console.error('Error finding book:', error);
-            res.status(500).json({ error: 'Failed to '});
+            res.status(500).json({ error: 'Failed to find wanted books'});
         };
     },
     async getOwnedBooks(req, res) {
-        const { user_id } = req.body;
+        const { user_id } = req.params.user_id;
         try {
             const owned_books = await db.query(
-                "SELECT \
-                    o.book_id, \
-                    b.isbn, \
-                    gbi.title, \
-                    gbi.author_name, \
-                    o.condition, \
-                    gbi.category_name, \
-                    gbi.publish_date, \
-                    gbi.language, \
-                    gbi.series_name \
-                FROM ownership o \
-                LEFT JOIN book b ON b.book_id = o.book_id \
-                LEFT JOIN LATERAL get_book_info(b.isbn) gbi ON TRUE \
-                WHERE o.user_id = ?",
+                `
+                SELECT
+                    o.book_id,
+                    b.isbn,
+                    gbi.title,
+                    gbi.author_name,
+                    o.condition,
+                    gbi.category_name,
+                    gbi.publish_date,
+                    gbi.language,
+                    gbi.series_name
+                FROM ownership o
+                LEFT JOIN book b ON b.book_id = o.book_id
+                LEFT JOIN LATERAL get_book_info(b.isbn) gbi ON TRUE
+                WHERE o.user_id = :user_id
+                `,
                 {
                     type: QueryTypes.SELECT,
-                    replacements: [ user_id ]
+                    replacements: {user_id: user_id}
                 }
             );
             res.status(200).json(owned_books);
         } catch (error) {
             console.error('Error finding book:', error);
-            res.status(500).json({ error: 'Failed to '});
+            res.status(500).json({ error: 'Failed to find owned books'});
         }
     },
     async addBookToWanted(req, res) {
         const { user_id, isbn, title, first_name, middle_name, last_name, category, publish_date, language, book_series } = req.body;
         try {
             const isbn_exists = await db.query(
-                "SELECT book_id FROM book WHERE isbn = ?",
+                "SELECT book_id FROM book WHERE isbn = :isbn",
             {
                 type: QueryTypes.SELECT,
-                replacements: [isbn] 
+                replacements: {isbn: isbn}
             });
             if (!isbn_exists) {
                 const jsonData = JSON.stringify([{
                     isbn, title, first_name, middle_name, last_name, category, publish_date, language, book_series
                 }])
-                await db.query("CALL add_books_bulk(?::JSONB)",
+                await db.query("CALL add_books_bulk(:json_data::JSONB)",
                     {
                         type: QueryTypes.RAW,
-                        replacements: [jsonData] 
+                        replacements: {json_data: jsonData}
                     }
                 )
             };
             const bookId = await db.query(
-                "SELECT book_id FROM book WHERE isbn = ?",
+                "SELECT book_id FROM book WHERE isbn = :isbn?",
                 {
                     type: QueryTypes.SELECT,
-                    replacements: [isbn]
+                    replacements: {isbn: isbn}
                 }
             );
             await db.query(
-                "INSERT INTO wanted (user_id, book_id) VALUES (?, ?)",
+                "INSERT INTO wanted (user_id, book_id) VALUES (:userid, :bookid)",
                 {
                     type: QueryTypes.INSERT,
-                    replacements: [user_id, bookId] 
+                    replacements: {userid: user_id, bookid: bookId}
                 }
             );
             res.status(200);
@@ -103,34 +106,34 @@ const BookController = {
         const { user_id, isbn, title, first_name, middle_name, last_name, category, publish_date, language, book_series, condition } = req.body;
         try {
             const isbn_exists = await db.query(
-                "SELECT book_id FROM book WHERE isbn = ?",
+                "SELECT book_id FROM book WHERE isbn = :isbn",
             {
                 type: QueryTypes.SELECT,
-                replacements: [isbn]
+                replacements: {isbn: isbn}
             });
             if (!isbn_exists) {
                 const jsonData = JSON.stringify([{
                     isbn, title, first_name, middle_name, last_name, category, publish_date, language, book_series
                 }])
-                await db.query("CALL add_books_bulk(?::JSONB)",
+                await db.query("CALL add_books_bulk(:json_data::JSONB)",
                     {
                         type: QueryTypes.RAW,
-                        replacements: [jsonData]
+                        replacements: {json_data: jsonData}
                     }
                 )
             };
             const bookId = await db.query(
-                "SELECT book_id FROM book WHERE isbn = ?",
+                "SELECT book_id FROM book WHERE isbn = :isbn",
                 {
                     type: QueryTypes.SELECT,
-                    replacements: [isbn]
+                    replacements: {isbn: isbn}
                 }
             );
             await db.query(
-                "INSERT INTO ownership (user_id, book_id, condition) VALUES (?, ?, ?)",
+                "INSERT INTO ownership (user_id, book_id, condition) VALUES (:user_id, :book_id, :condition)",
                 {
                     type: QueryTypes.INSERT,
-                    replacements: [user_id, bookId, condition]
+                    replacements: {user_id: user_id, book_id: bookId, condition: condition}
                 }
             );
             res.status(200);
@@ -143,21 +146,22 @@ const BookController = {
         const { user_id, isbn } = req.body;
         try {
             const book = await db.query(
-                "SELECT book_id FROM book WHERE isbn = ?",
+                "SELECT book_id FROM book WHERE isbn = :isbn",
                 {
                     type: QueryTypes.SELECT,
-                    replacements: [isbn]
+                    replacements: {isbn: isbn}
                 }
             );
+            console.log(book)
             if (!book) {
                 return res.status(404).json({error: "Book not found"});
             }
             const bookId = book[0].book_id;
             const result = await db.query(
-                "DELETE FROM wanted WHERE user_id = ? AND book_id = ?",
+                "DELETE FROM wanted WHERE user_id = :user_id AND book_id = :book_id",
                 {
                     type: QueryTypes.DELETE,
-                    replacements: [user_id, bookId]
+                    replacements: {user_id: user_id, book_id: bookId}
                 }
             );
             if (result.rowCount === 0){
@@ -175,10 +179,10 @@ const BookController = {
         const { user_id, isbn } = req.body;
         try {
             const book = await db.query(
-                "SELECT book_id FROM book WHERE isbn = ?",
+                "SELECT book_id FROM book WHERE isbn = :isbn",
                 {
                     type: QueryTypes.SELECT,
-                    replacements: [isbn]
+                    replacements: {isbn: isbn}
                 }
             );
             if (!book) {
@@ -186,10 +190,10 @@ const BookController = {
             }
             const bookId = book[0].book_id;
             const result = await db.query(
-                "DELETE FROM ownership WHERE user_id = ? AND book_id = ?",
+                "DELETE FROM ownership WHERE user_id = :user_id AND book_id = :book_id",
                 {
                     type: QueryTypes.DELETE,
-                    replacements: [user_id, bookId]
+                    replacements: {user_id: user_id, book_id: bookId}
                 }
             );
             if (result.rowCount === 0){
